Split up mocks.go by mocked service
